Create Digital Roots
Digital root is the recursive sum of all the digits in a number.

Given n, take the sum of the digits of n. If that value has more than one digit, continue reducing in this way until a single-digit number is produced. The input will be a non-negative integer.

ex) input = 16, output = 1+6 -> 7
input = 29, output = 11 -> 1+1 -> 2